nexus-llm-analytics-distribution_20251018_183430 (1)/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus-llm-analytics"
version = "1.0.0"
description = "Multi-Agent Data Analysis Assistant with local-first, privacy-preserving design"
authors = [{name = "KOTAHARSHA25", email = "your.email@example.com"}]
license = {file = "LICENSE"}
readme = "docs/README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "fastapi",
    "crewai",
    "crewai_tools",
    "langchain-community",
    "pandas",
    "polars",
    "pydantic",
    "uvicorn",
    "python-dotenv",
    "chromadb",
    "plotly",
    "PyPDF2",
    "reportlab",
    "RestrictedPython",
    "python-multipart",
    "werkzeug",
    "openpyxl",
    "seaborn",
    "scipy",
    "scikit-learn",
    "psutil",
]

[project.urls]
Homepage = "https://github.com/KOTAHARSHA25/nexus-llm-analytics"
Repository = "https://github.com/KOTAHARSHA25/nexus-llm-analytics"
Issues = "https://github.com/KOTAHARSHA25/nexus-llm-analytics/issues"

[project.scripts]
nexus-backend = "src.backend.main:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "e2e: End-to-end tests",
    "performance: Performance tests",
    "fuzz: Fuzz and property-based tests",
    "security: Security tests",
    "slow: Slow-running tests",
    "async: Async tests",
    "mock: Tests using mocks"
]

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | env
  | chroma_db
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true